update: implement logic to make new test pass

+++ src/report.py
@@ -1,3 +1,7 @@
+from os.path import exists
+
 def generate_report(file_path):
+    if not exists(file_path):
+        raise FileNotFoundError(f'no file found at {file_path}')
     with open('./output/report.json', 'w'):
         ...
